% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_error.R
\name{st_error}
\alias{st_error}
\title{Calculate standard error of Large sample}
\usage{
st_error(vect)
}
\arguments{
\item{vect}{numerical vector}
}
\value{
standard error
}
\description{
Calculate standard error of Large sample
}
\examples{
st_error(rnorm(50))
}
