% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_squares.R
\name{sum_squares}
\alias{sum_squares}
\title{Calculate sum squares for a sample}
\usage{
sum_squares(vect)
}
\arguments{
\item{vect}{numeric vector}
}
\value{
number
}
\description{
Calculates the sum of squared deviations from mean, SS(x) of a numeric vector
}
\examples{
sum_squares(rnorm(50))
}
